using double everywhere, throwing exception if iscore zero in derivatives, using weight for caching and log formatting

+++ src/main/java/com/kushalarora/compositionalLM/derivatives/dXdXw.java
@@ -40,8 +40,8 @@ public class dXdXw extends AbstractBaseDerivativeClass {
         considerCreatingArrays(length);
 
         INDArray[][] phraseMatrix = scorer.getPhraseMatrix();
-        float[][][] compositionISplitScore = scorer.getCompositionISplitScore();
-        float[][] compositionIScore = scorer.getInsideSpanProb();
+        double[][][] compositionISplitScore = scorer.getCompositionISplitScore();
+        double[][] compositionIScore = scorer.getInsideSpanProb();
 
 
         for (int i = 0; i < length; i++) {
